ixl: Add debug sysctl to dump internal switch data.
This sysctl's output can be written to a file and used as the input
for an adapter-specific switch debug application that can show
the MAC/VLAN filters that are in use.

It dumps out a lot of data, so I wouldn't recommend letting this
sysctl print straight to the console.

Signed-off-by: Eric Joyner <eric.joyner@intel.com>