ixl: Change constants used when getting FEC status
Previously, the driver used the constants defined for setting the
FEC bits in the Set PHY Config Admin Queue command for getting
the current FEC ability from Get PHY Abilities. Change that to use
the constants defined for the field in Get PHY Abilities.

In practice, this shouldn't change the output, but it could protect
against future Get PHY abilities response changes.

Signed-off-by: Eric Joyner <eric.joyner@intel.com>