virtchnl: add VIRTCHNL_OP_REQUEST_QUEUES
This patch adds a new virtchnl OP which will allow the VF to negotiate
with the PF how many queues the PF has allocated for that VF.  In the
current scheme each VF gets a default number of queues and the VF can
choose to enable any/all of those allocated queues.  This message is
intended to be used to request a differential number allocated queues
for that particular VF.  Without this new message there is no way for
the VF to request more queues than what it is allocated by default.

Signed-off-by: Alan Brady <alan.brady@intel.com>