ixl: Don't block speed changes on 25G SFP cards
Allow user to set a specific speed they want to advertise on 25G SFP28 cards.
Normally the LESM negotiates the best speed, but there is a timing issue that
could allow 10G speeds to be set on 25G media. So this should let the user
let the card only advertise 25G speeds.

Signed-off-by: Eric Joyner <eric.joyner@intel.com>