ixl: Fix kernel panic
When driver gets an error while getting PHY info in attach, it doesn't
create an ethernet interface. So on detach, the driver tries to remove
the interface that isn't there, causing a kernel panic.

Attach the interface anyway even if there is an error. It's possible
the adapter can provide the PHY abiliites information later.

Signed-off-by: Eric Joyner <eric.joyner@intel.com>