ixgbe(4): fix the FWSM.PT check in ixgbe_mng_present()
Previously we expected that only the pass-through bit will be set
in FWSM on MNG enabled NVM, however this is not the case and the
FLASH bit should always be enabled on a good NVM making the
current check invalid.

Change the logic to assume MNG is present only based on the
FWSM.PT bit.

Signed-off-by: Emil Tantilov <***@intel.com>