ixlv: properly free allocated filter lists
The filter lists for the ixlv driver are allocated memory, and not
static structs. At minimum, we need to free these pointers when we
de-initialize them. A bigger refactor might take the time to make them
static members of the ixlv_sc structure, instead of pointers, but this
requires more code refactor, as every reference must change to use the &
operator.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>