Revert "ixl: Add base_id to VF number used in VF RSS register macros"
It was mistakenly inferred that we need to use the global VF id when
writing registers to configure RSS, as it was a plausible explanation
why RSS would not work on the VFs of the 2nd physical port.

Based on empirical testing this is not the case, as RSS functions
correctly on Linux hosts, which do not program the registers using the
base_id. Additionally, testing on the ixl host driver shows that RSS is
still non-functional on VFs of the 2nd port.

Since the fix is not correct, revert it.

Reverts: f5e0c0159b09 ("ixl: Add base_id to VF number used in VF RSS register macros", 2016-12-14)
Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>