ixl: Remove interface re-init on watchdog event
Just calling ixl_init_locked() to reset the interface does not
seem to work to unhang a queue, and it also frequently causes other
errors like sends complaining about no buffer space available.

Due to this, remove the call to ixl_init_locked(), and instead
just log that a queue hang occurred and increment a stat. Since queue
hangs rarely occur and/or seem to be false positives, this should
suffice for now until we properly add support to safely reset
the queues and the driver.

Signed-off-by: Eric Joyner <eric.joyner@intel.com>