always bus_dmamap_destroy even when buf->map is zero
It turns out that the map "pointer" returned by bus_dmamap_create may be
zero. This resulted in a memory leak when unloading the device driver or
destroying a virtual function, as we would not call
bus_dmamap_destroy(), which in turn resulted in bus_dma_tag_destroy()
returning EBUSY. The end result was that about 24 M_DEVBUF allocations
were being leaked.

Fix this by removing checks for buf->map against NULL. These checks
are invalid, and not necessary. The actual pointer used to reference the
variable is the tag pointer, and not the value of the map variable.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>