ixlv: check AdminQ messages immediately during attach
The ixl PF driver was recently modified to notify the VF of current link
state right after receiving a GET_VF_RESOURCES message. Unfortunately,
the ixlv driver would not check the AdminQ until we received an
interrupt on its irq. Because the link state message is sent prior to
the AdminQ irq being setup, we do not see this message until another
message is sent by the PF, which likely won't happen until the device is
brought up via ifconfig.

Fix this by immediately queueing the AdminQ task as soon as we've
initialized the taskqueue. This ensures that we see the message and have
accurate link status as soon as possible.

Now that we expect to get a link status update, we also no longer need
to default to link_up.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>