ixl: use HW flag for NVM read lock check
We need the NVM lock for all reads on the X722 device, because
the read occurs over the AdminQ. Fix issues with NVM reads by implementing
a new I40E_HW_FLAG_* feature flag, which we enable always for X722 devices.

Add a firmware version check just after acquiring the NVM version
information, which we'll use to additionally enable the check for new
firmware versions.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>