ixl: re-use __i40e_read_nvm_word to reduce code duplication
We now have two implementations of i40e_read_nvm_word(), one which takes
the NVM lock if necessary, and one which assumes the NVM lock is already
held.

Instead, lets re-write i40e_read_nvm_word() in terms of
__i40e_read_nvm_word() so that we don't duplicate the code for selecting
which variant of the NVM read to use.

Additionally, slightly cleanup the control logic for handling the NVM
lock so that we return early if we can't obtain the NVM lock. This
reduces code maintenance upstream.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>