ixl: avoid acquiring the NVM lock twice in validate_nvm_checksum
Since we must always acquire the NVM lock here, lets use
__i40e_read_nvm_word from within the lock, so that we don't need to take
the lock multiple times. Also we can simplify the function by just
returning instead of using goto. This overall simplifies the logic, and
also makes sure that we don't attempt to release the NVM if we never
acquired it.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>