ixl: use reverse cleanup goto strategy on allocation failures
When setting up a queue, we currently rely on cleaning up using a single
"fail" label. Instead, split this into multiple labels unique to each
failure point which jump into the necessary section to clean up. This
ensures that we clean up everything in reverse and avoids needing to
rely on checking for initialization of each variable.

Additionally this pattern makes it more clear what needs to be free'd,
and ensures that we free it in reverse order. Otherwise we might
possibly free an outer structure before freeing the memory associated
with pointers stored within the structure.

As part of the refactor, now also cleanup the soft structures allocated
by ixl_allocate_tx_data. Previously this was overlooked and not handled
except in ixl_free_vsi(). This meant that callers of ixl_setup_queue had
to be aware of what parts they needed to cleanup after during an
allocation failure, lest they leak the memory.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>