ixlv: don't set que->res to NULL when bus_setup_intr() fails
When bus_setup_intr() fails, we set que->res to NULL. This is incorrect,
because que->res was successfully allocated moments prior. Setting it to
NULL now would cause us to never release the resources during cleanup.
Instead, we can rely on the cleanup flow to properly free the MSI-X
resource if que->res is non NULL.

We'll remove the que->res = NULL line, and replace it with que->tag =
NULL, since checking que->tag for NULL is how we decide whether to
release the interrupt.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>