ixlv: assign MSI-X interrupts prior to setting up interface
We assign the MSi-X queue interrupts after we've already setup the
interface. Howver, it is possible for this to fail, and we have no code
to cleanup and detach the created interface if we do so.

Instead of adding code to handle cleaning up the interface and ifmedia
setup, move this to the end after everything else that can fail. This
slightly delays when we attach the ether device, but means we no longer
need to worry about cleaning up incase of failure.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>