ixl/ixlv: Fix ixl_tso_detect_sparse()
An external customer reported MDD events with TSO, and provided
buffer lengths that passed the old detection function that should not
have. Fix the function so that these chains are properly detected as
needing defragmenting.

And increase the max chain size limit. The true limit is 7 due to an
errata with the header length being double counted; the six was in the
code because eight was wrong, but we weren't sure if seven was right.

Signed-off-by: Eric Joyner <eric.joyner@intel.com>