ixl: fix flag check in ixl_add_hw_filters
Currently, when we to add new MAC/VLAN filters to the hardware via the
AdminQ, we pass in a set of flags for which filters we're adding, and
then we check each filter to make sure the flags match.

This doesn't really work, since in ixl_reconfigure_filters, we only pass
in IXL_FILTER_UNUSED as the flag. This in theory was supposed to add all
filters marked as "UNUSED". However, this doesn't work properly because
the filters have more than just one flag bit set. This essentially means
that reconfigure_filters will never work the way we expect.

Fix this by having the ixl_add_hw_filters function first mask off any
bits we're not checking before comparison.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>