ixl: add sysctl to display the supported link speeds
Since we have a sysctl to set the advertised link speed, add a new
sysctl "supported_speeds" which shows the full set of supported link
speeds according to the same format. This is useful if the system
administrator wishes to know what speeds could be supported but are not
currently advertised.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>