ixl: initialize advertised_speed to supported_speeds
When the driver loads, the currently configured PHY advertised speeds is
reported by the sysctl. This causes some confusion, since if a user
changes this value, and reloads the driver (should it be compiled as
such), we'll effectively maintain configuration across driver loads.

Instead, lets use the supported_speeds value and reprogram the PHY at
driver load time to ensure that we always restore the advertised setting
back to the default of "all supported speeds enabled". This makes the
sysctl behave as others.

This may at first seem purely cosmetic, but it can cause some surprising
behaviors. Suppose a system administrator sets the value to 0 (or some
speed that is not supported by the link partner). Then, *even* after a
driver reload, the device will still keep the same speeds and could fail
to link.

This patch fixes this surprising behavior by restoring the default
supported set upon driver load.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>