ixl: avoid spurious unqualified module detected message
When link goes down, we check to see if the firmware indicated that the
module is not qualified. The following message is displayed to notify
the user of the reason for the link down event:

  ixl0: Link failed because an unqualified module was detected!

Unfortunately, if the user has set advertise_speed to zero to disable
link altogether, it causes the firmware to report all modules as
unqaulified. This results in displaying this message in a confusing
case.

Since we know that an advertised_speed of zero will cause modules to be
unqualified, avoid printing the message if the advertised_speed has been
set to zero.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>