ixl: fix incorrect SYSCTL_ADD_ULONG usage
These sysctls are actually of type u64, which means using
SYSCTL_ADD_ULONG is incorrect, as this will only work if unsigned long
happens to be 64bits.

Fix this by using SYSCTL_ADD_UQUAD, which will correctly add a uint64_t
sysctl.

This resolves compilation warnings similar to the following when on
32-bit platforms:

  ixl_pf_main.c:2837:2: error: incompatible pointer types initializing 'unsigned long *' with an expression of type 'u64 *' (aka 'unsigned long long *') [-Werror,-Wincompatible-pointer-types]
          SYSCTL_ADD_ULONG(ctx, child, OID_AUTO, "watchdog_events",
          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>