ixl: fix initialization of baudrate on FreeBSD 10
On FreeBSD prior to version 11, the ifp->baudrate is actually a ulong,
and not a u64 type. In order to store higher baud rates on 32-bit
platforms, FreeBSD 10 uses an extra baudrate_pf variable, which
indicates the power factor of the baud rate.

In order to allow initilization from a u64 type, the function
if_initbaudrate() was used. However, in FreeBSD 11, this was changed to
simply have ifp->baudrate be a uint64_t, while removing the function.

For versions of FreeBSD prior to 11, use if_initbaudrate in order to set
the baudrate correctly on those systems.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>