ixl(4): Hide weird flag behavior in TX CSO path
TCP checksums need to be calculated when TCP checksums or TSO is
requested by the stack, and this means the driver needs to compare against
more flags in the TCP case than in the UDP or SCTP case. So instead of having
a long value containing a bunch of OR'd flags, put it in a macro to make it
easier to read.

And put the UDP and SCTP values in macros, too, for consistency.

Signed-off-by: Eric Joyner <erj@freebsd.org>