ixl(4): Add driver version number to end of device description string
The nvmupdate tool checks the version number in the %desc sysctl in order to
run, so the version number needs to be visible there. To do so, this commit
creates a new macro PVIDV() that extends the existing PVID() by adding the
driver version string to the end of the device description string.

Signed-off-by: Eric Joyner <erj@freebsd.org>