iflib: Call pci_iov_detach() during iflib_device_deregister()
Currently, there's no way for a PF driver to destroy all of its VF
devices on detach, so the driver needs to exit detach and wait for
the VF devices to be destroyed by the user.

However, a combination of the driver-specific detach() function being called
too late in deregister with its return value not being checked makes
calling pci_iov_detach() in the driver a poor choice -- the detach process
will continue without the VFs being properly destroyed, causing issues
later.

So instead, mimic what was in the pre-iflib driver in iflib itself
-- the deregister process will bail if it sees that SR-IOV is still
in use for the device, and print a message telling the user to destroy
the VFs first.

Signed-off-by: Eric Joyner <erj@freebsd.org>