ixl(4): Enhance ixl_tso_detect_sparse() to fix more MDD issues
From Drew:

The problem with the current sparse detection logic is that it is "smarter than
the nic". e.g, your function accrues a "buffer" when the DMA segments are big,
and then sucks small DMA segments into that buffer.  This kind of "averages
out" the lengths, and you miss areas of the packet chain where packets could be
made up of many small segments.  The first packet below is an
example of such a case, and is similar to what was causing MDDs on Netflix
workloads.

Sample packet from above:

	pi.ipi_ehdrlen = 14;
	pi.ipi_ip_hlen = 20;
	pi.ipi_tcp_hlen = 32;
	pi.ipi_tso_segsz = 1448;

	segs[nseg++].ds_len = 66;
	segs[nseg++].ds_len = 13;
	segs[nseg++].ds_len = 4096;
	segs[nseg++].ds_len = 4096;
	segs[nseg++].ds_len = 4096;
	segs[nseg++].ds_len = 4096;
	segs[nseg++].ds_len = 16;
	segs[nseg++].ds_len = 13;
	segs[nseg++].ds_len = 4;
	segs[nseg++].ds_len = 16;
	segs[nseg++].ds_len = 13;
	segs[nseg++].ds_len = 4096;
	segs[nseg++].ds_len = 16;

(13 byte segment is from TLS)

Signed-off-by: Eric Joyner <erj@freebsd.org>