ixl/ixlv(4): Stop setting legacy intr field in iflib txrx function pointer struct directly
This structure is used by both the PF and VF driver, but only the PF driver
has legacy interrupt support, and the VF may not have access to the PF's
symbols.

So, set that field to NULL in ixl_txrx.c, and set the function pointer
in the PF driver only.

Signed-off-by: Eric Joyner <erj@freebsd.org>