iflib: Add public iflib_request_reset() for drivers to use
ixl(4) needs a way to tell iflib that the driver is going down and needs
to have its driver context re-initialized. It can now do that by setting the
IFC_DO_RESET flag in the iflib ctx via this new function, and that will have
iflib re-init the driver on the next _task_fn_admin() execution.

Signed-off-by: Eric Joyner <erj@freebsd.org>