ixlv(4): Add VF reset debug sysctls
There are two: debug.do_vf_reset and debug.do_vflr_reset; the first
sends a message to the PF asking it to reset the VF, and the second
writes a bit to PCIe config space requesting the hardware to reset
the VF (and that causes a VFLR interrupt to be sent to the PF, for it
to handle re-initializing the VF's resources).

Currently, the VF comes back up after a VF reset, but not after a
VFLR initiated by this sysctl.

Signed-off-by: Eric Joyner <erj@freebsd.org>