ixl(4): Lie to the VF about configuring promiscuous mode if VF isn't allowed to configure it
This mimics the behavior of the Linux PF driver, and will stop the VF from complaining
about errors when configuring promiscuous mode. There is a deeper problem here where
the VF doesn't know if it's allowed to configure promiscuous mode or not, but for now,
this stops the unnecessary error message printing.

Signed-off-by: Eric Joyner <erj@freebsd.org>