ixl(4): Add tunable to enable VF loopback; enable by default
This resolves the issue where VF-VF/VF-PF Ethernet traffic wasn't
working.

When SR-IOV is enabled and a VF is added, a virtual bridge is created
and both the PF and VF interfaces are placed behind it. Before this
patch, the bridge was always in VEPA mode and traffic could not be
routed directly between the interfaces connected to the bridge; it
had to go out to an external switch.

With this patch, we enable the PF and VF VSIs to allow loopback traffic,
which makes the virtual bridge act like a VEB, and allows traffic to
get switched between the PF/VF VSIs directly in the hardware.

Signed-off-by: Eric Joyner <erj@freebsd.org>