iflib: Prevent kernel panic on detach in _task_fn_admin by checking if in detach
The kernel panic above was caused by _task_fn_admin calling _sx_xlock_hard and
triggering a page fault. My theory is that a _task_fn_admin task was executing
during detach, and since the IFLIB_ADMIN_ALWAYS_RUN flag was set, it was ignoring
the !running state and trying to lock an sx lock that was destroyed.

Since iflib was already setting ifc_in_detach on detach (and before destroying
the ctx sx lock), this patch moves it out from the ownership of the ctx sx lock
and makes _task_fn_admin check it (atomically), so that it doesn't try to
lock the sx lock when the driver is in detach and the sx lock could be destroyed.

Signed-off-by: Eric Joyner <erj@freebsd.org>