ixl/ixlv(4): Fix panic when receiving jumbo frame (>4k mtu) traffic
It turns out that ixl_isc_rxd_available is supposed to return how many
packets are available to be cleaned on the rx ring. This patch removes
a section of code where if the budget argument is 1, the function would return
one if there was a descriptor available, not necessarily a packet.

This is okay in regular mtu 1500 traffic since the max frame size is less
than the configured receive buffer size (2048), but this doesn't work when
received packets can span more than one  descriptor, as is the case when the
mtu is 9000 and the receive buffer size is 4096.

Signed-off-by: Eric Joyner <erj@freebsd.org>