ixlv(4): Fix a memory leak when interrupt setup fails
When legacy setup fails, iflib_device_register calls iflib_tx_structures_free,
which then calls ixlv_if_queues_free. In that function, the RSQ for the
TX queues, when not in head writeback mode, is freed by looping over
the number of TX queues.

Before this patch, that loop was using a value set
in ixlv_if_attach_post, which hasn't been called yet when the legacy interrupt
setup failure occurs. So this patch changes the loop to use the tx queues value
stored in the iflib ctx.

Signed-off-by: Eric Joyner <erj@freebsd.org>