ixlv(4): Make ixlv_if_init() and ixlv_stop() wait for ENABLE/DISABLE queue command completion
This helps prevent a case where ifconfig up completes even though the PF hasn't processed the
corresponding ENABLE_QUEUES message, causing something like ARP to send a packet on a queue
that isn't enabled yet, causing a TX MDD event 0 to be triggered.

This needs to be tested when the driver is loaded at boot, and without EARLY_AP_STARTUP enabled.

Signed-off-by: Eric Joyner <erj@freebsd.org>