iflib: Move ifc_in_detach variable to flag in if_ctx_t's ifc_flags, and lock around it
The ifc_in_detach needs to be protected by a lock since it is written in one
thread (iflib_device_deregister) and read in other threads (e.g. _task_fn_admin),
Since the STATE_LOCK is already taken in _task_fn_admin and it appears to generally
protect the driver state, convert the ifc_in_detach value to a flag that's set
in ifc_flags, and read it along with the other flags in _task_fn_admin. As
well, take the STATE_LOCK when setting the value during iflib_device_deregister().

However, this conversion means a lock needs to be taken during
iflib_fl_bufs_free(), which I'm not sure is safe to do.

Signed-off-by: Eric Joyner <erj@freebsd.org>