ixl(4): Clear VF qtag queue flags when VF is being reinitialized
The PF keeps track of VF queue state (enabled/configured) using flags
in the qmgr qtag assigned to the function. These flags are supposed to be set
when queues are configured/enabled and unset when queues are disabled or
the VF is reset. They were not being cleared on reset, which could cause
the PF to print messages like "Queues are already enabled" when the VF driver
was being reloaded after an ungraceful detach e.g. when a VM force powers off.

Signed-off-by: Eric Joyner <erj@freebsd.org>